Split-up e2e test into test cases
We now use dedicated functions and files for the test cases to be able
keep the tests clean. Every test case expects a valid working operator
to be deployed and is expected to cleanup on its own.
