+++ test/tc_redeploy_operator_test.go
@@ -1,0 +1,33 @@
+// +build e2e
+
+/*
+Copyright 2020 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package e2e_test
+
+func (e *e2e) testCaseReDeployOperator([]string) {
+	// Clean up the operator
+	e.cleanupOperator(manifest)
+
+	// Deploy the operator again
+	e.deployOperator(manifest)
+}
+
+func (e *e2e) cleanupOperator(manifest string) {
+	// Clean up the operator
+	e.logf("Cleaning up operator")
+	e.kubectl("delete", "-f", manifest)
+}
